refactor storage module
