+++ tests/tests_storage/test_get_available_bike_list.py
@@ -1,5 +1,5 @@
 from app.models import Bike
-from app.storage import get_available_bike_list
+from app.storage.available_bike_list import get_available_bike_list
 
 
 async def test_get_available_bike_list_happy_path(fixture_prefilled_available_bike_list):
